requirements.txt
smolagents>=1.1.0
litellm
duckduckgo-search
ddgs
markdownify
youtube-transcript-api
openai-whisper
librosa
torch
torchvision
pillow
pandas
gradio
requests
python-dotenv